use sprintf
was encountering a weird bug when I used strcpy, not sure what it was?
this seems to work.

```
Invalid UTF-8 leading byte 0x000000a9 encountered when deserializing a UTF-8 string in wasm memory to a JS string!
```

^ this would show up in the JS console when I entered an 11-char string
into the text box, such as `aaaaaaaaaaa`. it only seemed to happen w/
len 11, which I discovered totally by accident. weird.